requirements/common.txt
flask
flask-sqlalchemy
pyjwt
python-dotenv
python-jose
